third_party/blink/web_tests/http/tests/websocket/send-after-close-on-unload-expected.txt

Test if WebSocket opened in iframe, referred in parent document is, and try to send() returns false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws on main document is ready.
PASS insert a iframe, where open ws called 'socket1'
PASS document.childWebSocket is non-null.
PASS 'socket1' is sent to the server. unload the iframe. expect receiving 'socket1' on ws...
PASS document.childWebSocket.send('send to closed socket') is undefined.
PASS closedSocket is "socket1"
PASS closeCode is "1001"
PASS document.childWebSocket.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
